@@ -1,22 +1,27 @@
 module github.com/adevinta/noe
 
-go 1.20
+go 1.21
 
 require (
 	github.com/abbot/go-http-auth v0.4.0
-	github.com/go-logr/logr v1.2.4
+	github.com/go-logr/logr v1.4.1
+	github.com/google/uuid v1.3.0
 	github.com/guseggert/pkggodev-client v0.0.0-20211029144512-2df8afe3ebe4
-	github.com/prometheus/client_golang v1.15.1
-	github.com/prometheus/client_model v0.4.0
-	github.com/sirupsen/logrus v1.9.2
+	github.com/pelletier/go-toml v1.9.5
+	github.com/prometheus/client_golang v1.18.0
+	github.com/prometheus/client_model v0.5.0
+	github.com/sirupsen/logrus v1.9.3
+	github.com/spf13/afero v1.9.5
 	github.com/stretchr/testify v1.8.4
-	golang.org/x/mod v0.10.0
-	gomodules.xyz/jsonpatch/v2 v2.3.0
-	k8s.io/api v0.26.3
-	k8s.io/apimachinery v0.27.2
-	k8s.io/client-go v0.26.3
-	k8s.io/utils v0.0.0-20230209194617-a36077c30491
-	sigs.k8s.io/controller-runtime v0.14.6
+	golang.org/x/mod v0.14.0
+	golang.org/x/tools v0.16.1
+	gomodules.xyz/jsonpatch/v2 v2.4.0
+	k8s.io/api v0.29.2
+	k8s.io/apimachinery v0.29.2
+	k8s.io/client-go v0.29.2
+	k8s.io/kubelet v0.29.2
+	k8s.io/utils v0.0.0-20230726121419-3b25d923346b
+	sigs.k8s.io/controller-runtime v0.17.2
 	sigs.k8s.io/e2e-framework v0.2.0
 )
 
