@@ -1,9 +1,11 @@
 package main
 
 import (
-	"context"
 	"flag"
+	"fmt"
 	"os"
+	"slices"
+	"strings"
 	"time"
 
 	"github.com/adevinta/noe/pkg/httputils"
@@ -14,38 +16,62 @@ import (
 	"github.com/adevinta/noe/pkg/registry"
 	"github.com/prometheus/client_golang/prometheus"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
+	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
 	"sigs.k8s.io/controller-runtime/pkg/manager/signals"
 	"sigs.k8s.io/controller-runtime/pkg/metrics"
+	metricsserver "sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/webhook"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
 
+var mainContext = signals.SetupSignalHandler()
+
 func main() {
-	var preferredArch, systemOS string
+	var preferredArch, schedulableArchs, systemOS string
 	var metricsAddr string
 	var registryProxies, matchNodeLabels string
+	var certDir string
+	var kubeletImageCredentialProviderBinBir, kubeletImageCredentialProviderConfig string
 
 	flag.StringVar(&preferredArch, "preferred-arch", "amd64", "Preferred architecture when placing pods")
+	flag.StringVar(&schedulableArchs, "cluster-schedulable-archs", "", "Comma separated list of architectures schedulable in the cluster")
 	flag.StringVar(&systemOS, "system-os", "linux", "Sole OS supported by the system")
 	flag.StringVar(&metricsAddr, "metrics-addr", ":8080", "The address the metric endpoint binds to.")
-	flag.StringVar(&metricsAddr, "registry-proxies", "", "Proxies to substitute in the registry URL in the form of docker.io=docker-proxy.company.corp,quay.io=quay-proxy.company.corp")
+	flag.StringVar(&certDir, "cert-dir", "./", "The directory where the TLS certificates are stored")
+	flag.StringVar(&registryProxies, "registry-proxies", "", "Proxies to substitute in the registry URL in the form of docker.io=docker-proxy.company.corp,quay.io=quay-proxy.company.corp")
 	flag.StringVar(&matchNodeLabels, "match-node-labels", "", "A set of pod label keys to match against node labels in the form of key1,key2")
+	flag.StringVar(&kubeletImageCredentialProviderBinBir, "image-credential-provider-bin-dir", "", "The path to the directory where credential provider plugin binaries are located.")
+	flag.StringVar(&kubeletImageCredentialProviderConfig, "image-credential-provider-config", "", "The path to the credential provider plugin config file.")
+
 	flag.Parse()
 
-	Main(signals.SetupSignalHandler(), "./", preferredArch, systemOS, metricsAddr, registryProxies, matchNodeLabels)
-}
+	var schedulableArchSlice []string
 
-func Main(ctx context.Context, certDir, preferredArch, systemOS, metricsAddr, registryProxies, matchNodeLabels string) {
+	if schedulableArchs != "" {
+		schedulableArchSlice = strings.Split(schedulableArchs, ",")
+	}
 
+	if preferredArch != "" && schedulableArchs != "" && !slices.Contains(schedulableArchSlice, preferredArch) {
+		err := fmt.Errorf("preferred architecture is not schedulable in the cluster")
+		log.DefaultLogger.WithError(err).Error("refusing to continue")
+		os.Exit(1)
+	}
+	ctrllog.SetLogger(log.NewLogr(log.DefaultLogger))
 	// Setup a Manager
-	log.DefaultLogger.WithContext(ctx).Println("setting up manager")
+	log.DefaultLogger.WithContext(mainContext).Println("setting up manager")
 	mgr, err := manager.New(config.GetConfigOrDie(), manager.Options{
-		MetricsBindAddress: metricsAddr,
-		Logger:             log.NewLogr(log.DefaultLogger),
+		Metrics: metricsserver.Options{
+			BindAddress: metricsAddr,
+		},
+		WebhookServer: webhook.NewServer(webhook.Options{
+			Port:    8443,
+			CertDir: certDir,
+		}),
+		Logger: log.NewLogr(log.DefaultLogger),
 	})
 	if err != nil {
-		log.DefaultLogger.WithContext(ctx).WithError(err).Error("unable to set up overall controller manager")
+		log.DefaultLogger.WithContext(mainContext).WithError(err).Error("unable to set up overall controller manager")
 		os.Exit(1)
 	}
 
@@ -59,8 +85,11 @@ func Main(ctx context.Context, certDir, preferredArch, systemOS, metricsAddr, re
 			},
 			registry.RegistryLabeller,
 		)),
+		registry.WithRegistryMetricRegistry(metrics.Registry),
+		registry.WithSchedulableArchitectures(schedulableArchSlice),
+		registry.WithAuthenticator(registry.NewAuthenticator(kubeletImageCredentialProviderConfig, kubeletImageCredentialProviderBinBir)),
 	)
-	containerRegistry = registry.NewCachedRegistry(containerRegistry, 1*time.Hour)
+	containerRegistry = registry.NewCachedRegistry(containerRegistry, 1*time.Hour, registry.WithCacheMetricsRegistry(metrics.Registry))
 
 	if err = controllers.NewPodReconciler(
 		"noe",
@@ -68,21 +97,15 @@ func Main(ctx context.Context, certDir, preferredArch, systemOS, metricsAddr, re
 		controllers.WithRegistry(containerRegistry),
 		controllers.WithMetricsRegistry(metrics.Registry),
 	).SetupWithManager(mgr); err != nil {
-		log.DefaultLogger.WithContext(ctx).WithError(err).Error("unable to create pod controller")
+		log.DefaultLogger.WithContext(mainContext).WithError(err).Error("unable to create pod controller")
 		os.Exit(1)
 	}
 
 	// Setup webhooks
-	log.DefaultLogger.WithContext(ctx).Println("setting up webhook server")
+	log.DefaultLogger.WithContext(mainContext).Println("setting up webhook server")
 	hookServer := mgr.GetWebhookServer()
 
-	hookServer.Port = 8443
-	hookServer.CertDir = certDir
-	decoder, err := admission.NewDecoder(mgr.GetScheme())
-	if err != nil {
-		log.DefaultLogger.WithContext(ctx).WithError(err).Error("unable to create admission decoder")
-		os.Exit(1)
-	}
+	decoder := admission.NewDecoder(mgr.GetScheme())
 
 	admissionHook := &webhook.Admission{
 		Handler: arch.NewHandler(
@@ -90,14 +113,14 @@ func Main(ctx context.Context, certDir, preferredArch, systemOS, metricsAddr, re
 			containerRegistry,
 			arch.WithMetricsRegistry(metrics.Registry),
 			arch.WithArchitecture(preferredArch),
+			arch.WithSchedulableArchitectures(schedulableArchSlice),
 			arch.WithOS(systemOS),
 			arch.WithDecoder(decoder),
 			arch.WithMatchNodeLabels(arch.ParseMatchNodeLabels(matchNodeLabels)),
 		),
 	}
-	admissionHook.InjectLogger(log.NewLogr(log.DefaultLogger))
 
-	log.DefaultLogger.WithContext(ctx).Println("registering webhooks to the webhook server")
+	log.DefaultLogger.WithContext(mainContext).Println("registering webhooks to the webhook server")
 	hookServer.Register(
 		"/mutate",
 		httputils.InstrumentHandler(
@@ -111,9 +134,9 @@ func Main(ctx context.Context, certDir, preferredArch, systemOS, metricsAddr, re
 		),
 	)
 
-	log.DefaultLogger.WithContext(ctx).Println("starting manager")
-	if err := mgr.Start(ctx); err != nil {
-		log.DefaultLogger.WithContext(ctx).WithError(err).Error("unable to run manager")
+	log.DefaultLogger.WithContext(mainContext).Println("starting manager")
+	if err := mgr.Start(mainContext); err != nil {
+		log.DefaultLogger.WithContext(mainContext).WithError(err).Error("unable to run manager")
 		os.Exit(1)
 	}
 }
